rename and add pool IDs
